GenAI/Task 3 AI Resume Screening System with Tracing/requirements.txt
langchain
langchain-groq
langsmith
python-dotenv